move store/v2/flat.* => store/v2/root.*
