+++ store/v2/root/root_store.go
@@ -1,4 +1,4 @@
-package flat
+package root
 
 import (
 	"errors"
