+++ store/v2/root/store.go
@@ -1,4 +1,4 @@
-package flat
+package root
 
 import (
 	"crypto/sha256"
