+++ store/v2/root/store_test.go
@@ -1,4 +1,4 @@
-package flat
+package root
 
 import (
 	"errors"
